feat: crafting preview cancellation
If you now press cancel, you will first request
the server for cancellation.
The server will track all the futures
and cancel them.
Once cancelled, the server will send a reply
to the client and the preview screen
will closed.

The client can force closed if they don't
want to wait.

Rationale:
If a task takes too long to complete,
the player becomes inpatient and closes the
screen. However, the task would still be
processing in the background even when closed.
This is not good, as this is using a thread
in the autocrafting pool and further requests
could take long to complete.
So, now we communicate with the server to
*actually* cancel them and make space in the
thread pool.