fix: crash when closing a wireless grid that became available after opening
This can happen if the network
can't resolve during opening,
but CAN resolve during closing.
It would try to remove the watcher
but it would fail because it
was never registered.
With this change, we always
register the watcher
even without network.

Fixes #1031

Closely related to #512
This fixes the crash condition
described in #512, but not
reloading the grid as soon
as it becomes available.