fix: network transmitter not reconnecting after network change on other side
A merge on the receiver side could override the network
on the receiver side, but not the transmitting side
(because it is one-way).
However, nodes would linger
in the graph of the transmitting
side.
Because of this, the transmitter
was not reconnecting with the
receiving side because the receiver node
is still in the graph, even
if it has another network.

This messes up the graph, and causes
problems and exceptions later down the line
(not further investigated.)

No clue why the watcher changes
were necessary, but after
some tests the grids still behave
correctly with merges.

Fixes #949