fix: crafting tree preview crash with same node properties
For example, if resource A is used 5 times
 EXACTLY in 2 separate branches,
this would throw a crash from the preview
widget "as the node already exists".
The cause of this is the node being
distinct, but yet the same because
they happen to have the same attributes.

This makes the node use a class
instead of a record so we rely on
default object memory hashing instead.