fix: losing track of losing resources that can be autocrafted
If you caused a network merge on the
network receiver side, the network loses
track of all patterns.

This was because of 5c43eccc, which was working
around a bug only fixed later in 8eab53fa.

I have verified that we don't receive multiple
autocrafting notifications anymore, when causing
merges AFTER a processing task has been started
and BEFORE it has completed. That means the bug
fixed in 5c43eccc hasn't been reintroduced.

Fixes #1158