fix: processing inserting ingredient types in the wrong order
The resources were already in the correct
order from autocrafting api as fixed
in 9703c76,
but got in the wrong order
if they have different types
because the sinks are iterated
in the order they are registered.

Sort the sinks first. Still not 100%
correct behavior but better than
it was before.

Fixes #952