perf: don't reload recipe constantly with remainder items
For example when crafting mystical agriculture
essences, the crystal can be reused across
crafts. However, because we momentarily
decrement the matrix slot to air,
the recipe would become invalid
and it would reload the recipe.

Now we mute the listeners while we
decrement and replace the item with the
remainder.

After that, the listeners are called.

Fixes #962